refactor: moved production CLI functions
