+++ src/cli/msig/index.ts
@@ -1,6 +1,6 @@
-import { PrivateKey, Name, Checksum256, PublicKey, Weight, NameType } from '@wharfkit/antelope';
+import { PrivateKey, Name, Checksum256, NameType } from '@wharfkit/antelope';
 import { EosioMsigContract, setSettings } from '../../sdk';
-import { ActionData, createSigner, getProducers } from '../../sdk/services/blockchain';
+import { ActionData, createSigner } from '../../sdk/services/blockchain';
 import settings from '../bootstrap/settings';
 import { govMigrate } from './govMigrate';
 import { newAccount } from './newAccount';
@@ -13,6 +13,7 @@ import { vestingBulk } from './vestingBulk';
 import { hyphaContractSet } from './hyphaContractSet';
 import { setResourceConfig } from './setResourceConfig';
 import { setBlockchainConfig } from './setBlockchainConfig';
+import { addProd, removeProd } from './producers';
 
 const eosioMsigContract = EosioMsigContract.Instance;
 
@@ -160,138 +161,27 @@ export default async function msig(args: string[]) {
                 { proposer, proposalName, privateKey, requested: newGovernanceAccounts, test }
             );
         } else if (proposalType === 'add-prod') {
-            const producer = '1.found.tmy';
-            const signingKey = PublicKey.from('EOS6A3TosyQZPa9g186tqVFa52AfLdkvaosy1XVEEgziuAyp5PMUj');
-
-            // fetch the existing schedule and their keys
-            const { pending, proposed, active } = await getProducers();
-
-            if (pending || proposed) throw new Error("Can't add a producer while there is a pending schedule");
-
-            if (active.producers.find((p) => p.producer_name.equals(producer)))
-                throw new Error('Producer already in the schedule');
-
-            const newSchedule = active.producers.map((p) => {
-                return {
-                    // eslint-disable-next-line camelcase
-                    producer_name: p.producer_name,
-                    authority: [
-                        'block_signing_authority_v0',
-                        {
-                            threshold: 1,
-                            keys: p.authority[1].keys.map((k) => {
-                                return {
-                                    key: k.key,
-                                    weight: k.weight,
-                                };
-                            }),
-                        },
-                    ],
-                };
-            });
-
-            newSchedule.push({
-                // eslint-disable-next-line camelcase
-                producer_name: Name.from(producer),
-                authority: [
-                    'block_signing_authority_v0',
-                    {
-                        threshold: 1,
-                        keys: [
-                            {
-                                key: signingKey,
-                                weight: Weight.from(1),
-                            },
-                        ],
-                    },
-                ],
-            });
-            const action = {
-                account: 'tonomy',
-                name: 'setprods',
-                authorization: [
-                    {
-                        actor: 'tonomy',
-                        permission: 'owner',
-                    },
-                    {
-                        actor: 'tonomy',
-                        permission: 'active',
-                    },
-                ],
-                data: {
-                    schedule: newSchedule,
-                },
-            };
-
-            const proposalHash = await createProposal(
-                proposer,
-                proposalName,
-                [action],
-                privateKey,
-                newGovernanceAccounts
+            addProd(
+                {},
+                {
+                    proposer,
+                    proposalName,
+                    privateKey,
+                    requested: newGovernanceAccounts,
+                    test,
+                }
             );
-
-            if (test) await executeProposal(proposer, proposalName, proposalHash);
         } else if (proposalType === 'remove-prod') {
-            const producer = '1.found.tmy';
-
-            // fetch the existing schedule and their keys
-            const { pending, proposed, active } = await getProducers();
-
-            if (pending || proposed) throw new Error("Can't remove a producer while there is a pending schedule");
-
-            if (!active.producers.find((p) => p.producer_name.equals(producer)))
-                throw new Error('Producer not in the schedule');
-
-            const newSchedule = active.producers
-                .filter((p) => !p.producer_name.equals(producer))
-                .map((p) => {
-                    return {
-                        // eslint-disable-next-line camelcase
-                        producer_name: p.producer_name,
-                        authority: [
-                            'block_signing_authority_v0',
-                            {
-                                threshold: 1,
-                                keys: p.authority[1].keys.map((k) => {
-                                    return {
-                                        key: k.key,
-                                        weight: k.weight,
-                                    };
-                                }),
-                            },
-                        ],
-                    };
-                });
-
-            const action = {
-                account: 'tonomy',
-                name: 'setprods',
-                authorization: [
-                    {
-                        actor: 'tonomy',
-                        permission: 'owner',
-                    },
-                    {
-                        actor: 'tonomy',
-                        permission: 'active',
-                    },
-                ],
-                data: {
-                    schedule: newSchedule,
-                },
-            };
-
-            const proposalHash = await createProposal(
-                proposer,
-                proposalName,
-                [action],
-                privateKey,
-                newGovernanceAccounts
+            await removeProd(
+                {},
+                {
+                    proposer,
+                    proposalName,
+                    privateKey,
+                    requested: newGovernanceAccounts,
+                    test,
+                }
             );
-
-            if (test) await executeProposal(proposer, proposalName, proposalHash);
         } else if (proposalType === 'hypha-contract-set') {
             await hyphaContractSet(
                 {},
